Separate into modules and fix clippy warnings
