+++ rust/crates/cove_nfc/src/lib.rs
@@ -1,4 +1,8 @@
+pub mod header;
+pub mod ndef_type;
 pub mod parser;
+pub mod payload;
+pub mod record;
 
 pub struct NfcReader {}
