+++ rust/crates/cove_nfc/src/ndef_type.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum NdefType {
+    Empty,
+    WellKnown,
+    Mime,
+    AbsoluteUri,
+    External,
+    Unknown,
+    Unchanged,
+    Reserved,
+}
